# Conflicts:
#	jsettlers.logic/src/main/java/jsettlers/ai/highlevel/pioneers/PioneerGroup.java
#	jsettlers.logic/src/main/java/jsettlers/logic/map/grid/MainGrid.java
#	jsettlers.logic/src/main/java/jsettlers/logic/movable/Movable.java

+++ jsettlers.logic/src/main/java/jsettlers/algorithms/construction/AbstractConstructionMarkableMap.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015
+ * Copyright (c) 2015 - 2017
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -77,7 +77,7 @@ public abstract class AbstractConstructionMarkableMap {
 
 	public abstract boolean isInBounds(int x, int y);
 
-	public abstract boolean canConstructAt(short x, short y, EBuildingType type, byte playerId);
+	public abstract boolean canConstructAt(int x, int y, EBuildingType type, byte playerId);
 
 	public abstract byte calculateConstructionMarkValue(int mapX, int mapY, final RelativePoint[] flattenPositions);
 }
